+++ t-sparqlConstruct/src/main/java/eu/unifiedviews/plugins/transformer/sparql/construct/SparqlConstruct.java
@@ -6,6 +6,7 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.regex.Pattern;
 
+import eu.unifiedviews.helpers.dpu.extension.rdf.profiler.RdfProfiler;
 import org.eclipse.rdf4j.model.IRI;
 import org.eclipse.rdf4j.query.MalformedQueryException;
 import org.eclipse.rdf4j.query.QueryLanguage;
@@ -59,6 +60,9 @@ public class SparqlConstruct extends AbstractDpu<SparqlConstructConfig_V1> {
 
     @ExtensionInitializer.Init
     public RdfValidation rdfValidation;
+
+    @ExtensionInitializer.Init
+    public RdfProfiler rdfProfiler;
 
     @ExtensionInitializer.Init
     public FaultTolerance faultTolerance;