+++ e-httpRequest/CHANGELOG.md
@@ -1,6 +1,12 @@
 E-HttpRequest
 ----------
 
+v2.1.0
+---
+* Support for HTTP PUT/DELETE added
+* Accept also empty list of files in case of HTTP POST, do not fail because of that
+* Fixed encoding in case of multipart form HTTP POST request (#422)
+
 v2.0.0
 ---
 * Updated to use UnifiedViews helpers 3.X with rdf4j support