+++ e-sparqlEndpoint/CHANGELOG.md
@@ -1,6 +1,10 @@
 E-SparqlEndpoint
 ----------
 
+v2.0.1
+---
+* Added support for [RDF Profiler extension](https://grips.semantic-web.at/display/UDDOC/RDF+Profiler)
+
 v2.0.0
 ---
 * Connections properly closed