+++ t-rdfGraphMerger/CHANGELOG.md
@@ -1,6 +1,11 @@
 T-RdfGraphMerger
 ----------
 
+v2.0.1
+---
+* Added support for [RDF Profiler extension](https://grips.semantic-web.at/display/UDDOC/RDF+Profiler)
+
+
 v2.0.0
 ---
 * Updated to use UnifiedViews helpers 3.X with rdf4j support