+++ t-rdfToFiles/CHANGELOG.md
@@ -1,6 +1,10 @@
 T-RdfToFiles
 ----------
 
+v3.0.1
+---
+* Handling of RDFFormats fixed
+
 v3.0.0
 ---
 * Updated to use UnifiedViews helpers 3.X with rdf4j support