+++ t-relational/CHANGELOG.md
@@ -1,6 +1,10 @@
 T-Relational
 ----------
 
+v3.0.1
+---
+* Handling of connections adjusted
+
 v3.0.0
 ---
 * Updated to use UnifiedViews helpers 3.X with rdf4j support